backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
ultralytics==8.1.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
